Project 2 - 10/15/2018 (UI Updates)
Updated the CPU and PCB displays. Added functionality for Memory display to actually display what is in main memory, and for CPU display to display CPU status. Working now on PCB display to show what is on the ready queue. This is the kind of stuff that I hate, I'm not a fan of UI design.